I'm going to use SIMD for the traversal of nbrs of each node in BFS of builing mRR-set.

+++ code/Argument.h
@@ -10,7 +10,7 @@ using namespace std;
 
 // global variables
 Graph R_graph, O_graph;  
-vector<double> Inv_inDeg;
+vector<float> Inv_inDeg;
 int __numV_left;
 double __eta_left;
 int root_num;
